move util interp into viz
